# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SkyCheck Digital Check-In API
  description: |
    Backend API for SkyHigh Airlines digital check-in system.
    
    **Key Features:**
    - Time-bound seat reservations (120 seconds)
    - Conflict-free seat assignment with optimistic locking
    - Automatic hold expiration
    - Baggage validation with payment processing
    - High-performance seat map (P95 < 1s)

  version: 1.0.0
  contact:
    name: SkyHigh Airlines Engineering
    email: engineering@skyhigh.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/flights/{flightId}:
    get:
      tags:
        - Flights
      summary: Get flight by ID
      description: Retrieve flight details by flight ID
      operationId: getFlightById
      parameters:
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Flight retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'
        '404':
          description: Flight not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/flights/{flightId}/seats:
    get:
      tags:
        - Seat Map
      summary: Get seat map for flight
      description: |
        Retrieve all seats for a flight with current availability status.
        **Performance:** Cached for 10 seconds, P95 < 1000ms
      operationId: getSeatMap
      parameters:
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Seat map retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatMapResponse'
        '404':
          description: Flight not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/flights/{flightId}/seats/available:
    get:
      tags:
        - Seat Map
      summary: Get available seats
      description: Retrieve only available seats for a flight
      operationId: getAvailableSeats
      parameters:
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Available seats retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeatDto'
...
components:
  schemas:
    Flight:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        flightNumber:
          type: string
          example: SH101
        departureTime:
          type: string
          format: date-time
          example: '2026-03-08T14:00:00'
        arrivalTime:
          type: string
          format: date-time
          example: '2026-03-08T17:00:00'
        origin:
          type: string
          example: JFK
        destination:
          type: string
          example: LAX
        aircraftType:
          type: string
          example: Boeing 737
        status:
          type: string
          enum: [SCHEDULED, BOARDING, DEPARTED, ARRIVED, CANCELLED, DELAYED]
          example: SCHEDULED
    SeatDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        seatNumber:
          type: string
          example: 12A
        seatType:
          type: string
          enum: [WINDOW, MIDDLE, AISLE]
          example: WINDOW
        seatClass:
          type: string
          enum: [ECONOMY, BUSINESS, FIRST_CLASS]
          example: ECONOMY
        state:
          type: string
          enum: [AVAILABLE, HELD, CONFIRMED]
          example: AVAILABLE
        flightId:
          type: integer
          format: int64
          example: 1
    SeatMapResponse:
      type: object
      properties:
        flightId:
          type: integer
          format: int64
          example: 1
        flightNumber:
          type: string
          example: SH101
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatDto'
        totalSeats:
          type: integer
          example: 30
        availableSeats:
          type: integer
          example: 25
        heldSeats:
          type: integer
          example: 3
        confirmedSeats:
          type: integer
          example: 2
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-03-08T14:00:00'
        status:
          type: integer
          example: 409
        error:
          type: string
          example: Seat Already Reserved
        message:
          type: string
          example: Seat 12A is not available. Current state HELD
        path:
          type: string
          example: /api/check-in/101/hold-seat
        validationErrors:
          type: object
          additionalProperties:
            type: string
          nullable: true